+++ src/app/about/page.tsx
@@ -1,35 +1,25 @@
 'use client';
 import { Metadata } from 'next';
-import { useRouter } from 'next/navigation';
 import React from 'react';
 
 import { Info } from '@/widgets/company/Info';
 
-import { Breadcrumbs } from '@/features/Breadcrumbs/Breadcrumbs';
+import { Breadcrumbs } from '@/features/Breadcrumbs';
 
 export const metadata: Metadata = {
   title: 'О компании',
 };
 
+const breadcrumbs = [
+  { text: 'Главная', url: '/' },
+  { text: 'О компании', url: '/about' },
+];
+
 const About = () => {
-  const router = useRouter();
-
-  const goToNews = () => {
-    router.push('/about/news');
-  };
-
-  const breadcrumbs = [
-    { text: 'Главная', url: '/' },
-    { text: 'О компании', url: '/about' },
-  ];
-
   return (
     <>
       <Breadcrumbs breadcrumbs={breadcrumbs} />
       <Info />
-
-      {/* Используем функцию goToNews для кнопки */}
-      <button onClick={goToNews}>Перейти к новостям</button>
     </>
   );
 };
