+++ src/app/catalog/page.tsx
@@ -1,11 +1,21 @@
 import { Metadata } from 'next';
 
+import { Breadcrumbs } from '@/features/Breadcrumbs';
 export const metadata: Metadata = {
   title: 'Catalog',
 };
 
+const breadcrumbs = [
+  { text: 'Главная', url: '/' },
+  { text: 'Каталог', url: '/catalog' },
+];
+
 const Catalog = () => {
-  return <h1>Catalog</h1>;
+  return (
+    <>
+      <Breadcrumbs breadcrumbs={breadcrumbs} />
+    </>
+  );
 };
 
 export default Catalog;
