+++ src/features/Breadcrumbs/index.ts
@@ -1,0 +1,3 @@
+'use client';
+
+export { Breadcrumbs } from './ui';
