+++ src/features/Breadcrumbs/typing.ts
@@ -1,0 +1,6 @@
+export type Tbreadcrumb = {
+  breadcrumbs: {
+    text: string;
+    url: string;
+  }[];
+};
