+++ tsconfig.json
@@ -36,7 +36,7 @@
     ".next/types/**/*.ts",
     "src/features/Breadcrumbs/BreadcrumbItem.tsx",
     "src/features/Breadcrumbs/BreadcrumbManager.tsx",
-    "src/features/Breadcrumbs/Breadcrumbs.tsx"
+    "src/features/Breadcrumbs/ui.tsx"
   ],
   "exclude": ["node_modules", ".next", ".storybook"]
 }
